certvalidator/backend/requirements.txt
# API framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Async task queue
celery==5.3.6
redis==5.0.4

# Database
sqlalchemy==2.0.29
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# ML inference (shared with ml/)
torch==2.2.2
torchvision==0.17.2
transformers==4.40.0
onnxruntime-gpu==1.17.3
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4

# PDF
PyMuPDF==1.24.1
pdf2image==1.17.0

# Utilities
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
rich==13.7.1
